Whitelist system for reddit responses

+++ reddit.py
@@ -5,20 +5,25 @@ import traceback
 
 reddit = None
 log = None
+whitelist = None
 
 
-def init(logger):
+def init(logger, responseWhitelist):
 	global reddit
 	global log
+	global whitelist
 
 	reddit = praw.Reddit(user_agent=globals.USER_AGENT, log_request=0)
 	OAuth = OAuth2Util.OAuth2Util(reddit)
 	OAuth.refresh(force=True)
 
 	log = logger
+	whitelist = responseWhitelist
 
 
 def sendMessage(recipient, subject, message):
+	if whitelist is not None and recipient not in whitelist:
+		return True
 	try:
 		reddit.send_message(
 			recipient=recipient,
@@ -36,10 +41,13 @@ def getMessages():
 
 
 def markMessageRead(message):
-	message.mark_as_read()
+	if whitelist is None:
+		message.mark_as_read()
 
 
 def replyMessage(message, body):
+	if whitelist is not None and message.author not in whitelist:
+		return True
 	try:
 		message.reply(body)
 		return True
@@ -53,6 +61,8 @@ def getSubmission(id):
 
 
 def deleteComment(id=None, comment=None):
+	if whitelist is not None and not len(whitelist):
+		return True
 	try:
 		if id is not None:
 			reddit.get_info(thing_id='t1_' + id).delete()
@@ -66,7 +76,10 @@ def deleteComment(id=None, comment=None):
 
 def replyComment(id, message):
 	try:
-		resultComment = reddit.get_info(thing_id='t1_' + id).reply(message)
+		parent = reddit.get_info(thing_id='t1_' + id)
+		if whitelist is not None and parent.author not in whitelist:
+			return None
+		resultComment = parent.reply(message)
 		return resultComment.id
 	except Exception as err:
 		log.warning(traceback.format_exc())
@@ -74,6 +87,8 @@ def replyComment(id, message):
 
 
 def editComment(id, message):
+	if whitelist is not None and not len(whitelist):
+		return True
 	try:
 		reddit.get_info(thing_id='t1_' + id).edit(message)
 		return True
